+++ RTv1/Makefile
@@ -4,7 +4,7 @@ GCC = gcc
 
 CFLAGS = -Wall -Wextra -Werror
 
-OBJ = obj/main.o obj/parser.o obj/raytracer.o obj/draw_gest_scene.o obj/key.o obj/draw_tools.o obj/draw_value.o obj/key_see_obj.o obj/init_new_obj.o
+OBJ = obj/main.o obj/parser.o obj/raytracer.o obj/draw_gest_scene.o obj/key.o obj/draw_tools.o obj/draw_value.o obj/key_see_obj.o obj/init_new_obj.o obj/key_add_num.o obj/key_add_to_struct.o obj/draw_value_obj.o obj/draw_new_obj.o
 
 LIBFT = ./libft/libft.a
