+++ RTv1/example.scene
@@ -1,2 +1,1 @@
-S position{80, -10, -150} radius{80}
-D position{10, 10, 10} radius{10}
+D position{-1, -15, -20} Normal{3, 10, -10} radius{43}
