+++ RTv1/src/draw_gest_scene.c
@@ -51,49 +51,6 @@ static void		draw_list_obj(t_env *env)
 	draw_value(env);
 }
 
-static void		draw_new_obj_value(t_env *env)
-{
-	if (ft_strcmp(env->control->name_obj, "Sphere") == 0 ||
-		ft_strcmp(env->control->name_obj, "Disk") == 0)
-	{
-		mlx_string_put(env->mlx, env->win_control, 70, 200, 0x0000000, "Xc : ");
-		mlx_string_put(env->mlx, env->win_control, 70, 225, 0x0000000, "Yc : ");
-		mlx_string_put(env->mlx, env->win_control, 70, 250, 0x0000000, "Zc : ");
-		mlx_string_put(env->mlx, env->win_control, 70, 275, 0x0000000, "Radius : ");
-		if (ft_strcmp(env->control->name_obj, "Sphere") == 0)
-			draw_sphere_value(env);
-		else
-			draw_disk_value(env);
-	}
-	if (ft_strcmp(env->control->name_obj, "Plane") == 0)
-	{
-		mlx_string_put(env->mlx, env->win_control, 70, 200, 0x0000000, "Xc : ");
-		mlx_string_put(env->mlx, env->win_control, 70, 225, 0x0000000, "Yc : ");
-		mlx_string_put(env->mlx, env->win_control, 70, 250, 0x0000000, "Zc : ");
-		draw_plane_value(env);
-	}
-}
-
-static void		draw_new_obj_add(t_env *env)
-{
-	draw_square(env);
-	mlx_string_put(env->mlx, env->win_control, 100, 25, 0xFFFFFFF, "New Object");
-	mlx_string_put(env->mlx, env->win_control, 80, 150, 0x0000000, "Type : ");
-	mlx_string_put(env->mlx, env->win_control, 150, 150, 0x00000FF, env->control->name_obj);
-	mlx_string_put(env->mlx, env->win_control, 60, 400, 0x7A4500, "'Enter' to valided.");
-	draw_new_obj_value(env);
-}
-
-static void		draw_new_obj_menu(t_env *env)
-{
-	mlx_string_put(env->mlx, env->win_control, 90, 25, 0xFFFFFFF, "New Object");
-	mlx_string_put(env->mlx, env->win_control, 30, 100, 0xFFFFFFF, "Choose : ");
-	mlx_string_put(env->mlx, env->win_control, 35, 150, 0xFFFFFFF, " Sphere");
-	mlx_string_put(env->mlx, env->win_control, 35, 175, 0xFFFFFFF, " Plane");
-	mlx_string_put(env->mlx, env->win_control, 35, 200, 0xFFFFFFF, " Disk");
-	ft_draw_small_arrow(env);
-}
-
 void			draw_control(t_env *env)
 {
 	mlx_clear_window(env->mlx, env->win_control);
