+++ RTv1/src/key.c
@@ -1,5 +1,41 @@
 #include "../rtv1.h"
 
+static int		check_length(t_env *env)
+{
+	int 	ret;
+
+	ret = 0;
+	if (env->control->ctm_1 == 0)
+		ret = env->sphere->length == 0 ? 0 : 1;
+	if (env->control->ctm_1 == 1)
+		ret = env->plane->length == 0 ? 0 : 1;
+	if (env->control->ctm_1 == 2)
+		ret = env->disk->length == 0 ? 0 : 1;
+	return (ret);
+}
+
+static int		key_obj_menu(int keycode, t_env *env)
+{
+	if (keycode == KEY_ESC)
+	{
+		env->control->menu = 0;
+		env->control->ctm_1 = 0;
+	}
+	if (keycode == KEY_DOWN && env->control->ctm_1 < 2)
+		env->control->ctm_1++;
+	if (keycode == KEY_UP && env->control->ctm_1 > 0)
+		env->control->ctm_1--;
+	if (keycode == KEY_ENTER)
+	{
+		if (!check_length(env))
+			return (0);
+		env->control->name_obj = search_type_obj(env->control->ctm_1);
+		env->control->menu = 10;
+		env->control->ctm_1 = 0;
+	}
+	return (0);
+}
+
 static int		key_main_menu(int keycode, t_env *env)
 {
 	if (keycode == KEY_ESC)
@@ -51,7 +87,5 @@ int				key_hook(int keycode, t_env *env)
 		key_obj_in(keycode, env);
 	else if (env->control->menu == 20)
 		key_new_obj_in(keycode, env);
-	//ft_putnbr(keycode);
-	//ft_putendl("");
 	return (0);
 }
